test(ai): split planning_helpers_test to satisfy non-comment line limit (#3717)
planning_helpers_test.dart reached 1018 non-comment lines, tripping the
repo.dart_file_non_comment_line_size gate (limit 1000) in the Quality
workflow. Move the four invadable-province ownership helper groups into a
sibling planning_helpers_part2_test.dart so both files stay under the
limit. Pure test relocation; no production code change.
