+++ packages/colonizethis_ai/test/planning/planning_helpers_test.dart
@@ -21,9 +21,12 @@
 //     history scan, strict `<` cooldown window, predicate filtering (Refs #3717)
 //   - `atWarPeaceTargetBonus` — at-war GP eligibility gate, lazy predicate
 //     short-circuit, flat-bonus emission (Refs #3717)
-//   - `factionOwnsInvadableOldWorldProvince` — single-faction invadable-frontier
-//     ownership scan, `.any` short-circuit, absent/other-owner exclusion
-//     (Refs #3717)
+//
+// Invadable-province ownership helpers (`anyInvadableProvinceOwnedByMinor`,
+// `anyInvadableProvinceOwnedByGreatPower`, `factionOwnsInvadableOldWorldProvince`,
+// `addInvadableProvinceMinorOwnersNotAtWar`) are pinned in the sibling
+// `planning_helpers_part2_test.dart` to keep each file within the repo
+// non-comment line limit.
 
 import 'package:colonizethis_ai/src/perception/perception_snapshot.dart';
 import 'package:colonizethis_ai/src/planning/observer_goal_phase.dart';
@@ -38,8 +41,6 @@ import 'package:colonizethis_data/colonizethis_data.dart'
         isStalledOldWorldExpansion,
         kObserverConquestMinOwProvincesPerGp,
         kStalledOldWorldProvinceThreshold;
-import 'package:colonizethis_ai/src/util/faction_query.dart'
-    show isMinorFaction;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
@@ -94,17 +95,6 @@ AIWorldSnapshot _snapshotWithOwnOw(int oldWorldProvincesOwned) {
     threats: const ThreatSummary(atWarWith: []),
     opportunities: const OpportunitySummary(),
     conquest: ConquestSummary(oldWorldProvincesOwned: oldWorldProvincesOwned),
-    economy: const EconomySummary(),
-    relations: const {},
-  );
-}
-
-AIWorldSnapshot _snapshotWithInvadable(List<String> invadableProvinceIds) {
-  return AIWorldSnapshot(
-    playerId: _gp1,
-    threats: const ThreatSummary(atWarWith: []),
-    opportunities: const OpportunitySummary(),
-    conquest: ConquestSummary(invadableProvinceIdsSorted: invadableProvinceIds),
     economy: const EconomySummary(),
     relations: const {},
   );
@@ -786,371 +776,4 @@ void main() {
     });
   });
 
-  group('anyInvadableProvinceOwnedByMinor (Refs #3717)', () {
-    const String pA = 'provA';
-    const String pB = 'provB';
-
-    test('true when an invadable province is owned by a minor nation', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        anyInvadableProvinceOwnedByMinor(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _minor1},
-        ),
-        isTrue,
-      );
-    });
-
-    test('false when invadable provinces are owned only by GPs / tribes', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      expect(
-        anyInvadableProvinceOwnedByMinor(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2, pB: _tribe1},
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when an invadable province owner is absent from the map', () {
-      // Unowned / not-yet-mapped invadable province: lookup is null, no match.
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        anyInvadableProvinceOwnedByMinor(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {},
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when there are no invadable provinces', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable(const []);
-      expect(
-        anyInvadableProvinceOwnedByMinor(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _minor1},
-        ),
-        isFalse,
-      );
-    });
-
-    test('true when only a non-first invadable province is minor-owned', () {
-      // The .any short-circuit must still find a minor owner that is not the
-      // first scanned entry (GP first, minor second) -> deterministic true.
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      expect(
-        anyInvadableProvinceOwnedByMinor(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2, pB: _minor1},
-        ),
-        isTrue,
-      );
-    });
-  });
-
-  group('anyInvadableProvinceOwnedByGreatPower (Refs #3717)', () {
-    const String pA = 'provA';
-    const String pB = 'provB';
-
-    test('true when an invadable province is owned by a Great Power', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        anyInvadableProvinceOwnedByGreatPower(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2},
-        ),
-        isTrue,
-      );
-    });
-
-    test(
-      'false when invadable provinces are owned only by minors / tribes',
-      () {
-        final game = _gameWithGps();
-        final snapshot = _snapshotWithInvadable([pA, pB]);
-        expect(
-          anyInvadableProvinceOwnedByGreatPower(
-            game: game,
-            snapshot: snapshot,
-            provinceOwner: const {pA: _minor1, pB: _tribe1},
-          ),
-          isFalse,
-        );
-      },
-    );
-
-    test('false when an invadable province owner is absent from the map', () {
-      // Unowned / not-yet-mapped invadable province: `?? ''` -> playerById null.
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        anyInvadableProvinceOwnedByGreatPower(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {},
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when there are no invadable provinces', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable(const []);
-      expect(
-        anyInvadableProvinceOwnedByGreatPower(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2},
-        ),
-        isFalse,
-      );
-    });
-
-    test('true when only a non-first invadable province is GP-owned', () {
-      // The .any short-circuit must still find a GP owner that is not the
-      // first scanned entry (minor first, GP second) -> deterministic true.
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      expect(
-        anyInvadableProvinceOwnedByGreatPower(
-          game: game,
-          snapshot: snapshot,
-          provinceOwner: const {pA: _minor1, pB: _gp2},
-        ),
-        isTrue,
-      );
-    });
-  });
-
-  group('factionOwnsInvadableOldWorldProvince (Refs #3717)', () {
-    const String pA = 'provA';
-    const String pB = 'provB';
-
-    test('true when the faction owns an invadable province', () {
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        factionOwnsInvadableOldWorldProvince(
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2},
-          factionId: _gp2,
-        ),
-        isTrue,
-      );
-    });
-
-    test('false when invadable provinces are owned by other factions', () {
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      expect(
-        factionOwnsInvadableOldWorldProvince(
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp3, pB: _minor1},
-          factionId: _gp2,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when the owner lookup is absent from the map', () {
-      // Unowned / not-yet-mapped invadable province: lookup is null, no match.
-      final snapshot = _snapshotWithInvadable([pA]);
-      expect(
-        factionOwnsInvadableOldWorldProvince(
-          snapshot: snapshot,
-          provinceOwner: const {},
-          factionId: _gp2,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when there are no invadable provinces', () {
-      final snapshot = _snapshotWithInvadable(const []);
-      expect(
-        factionOwnsInvadableOldWorldProvince(
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp2},
-          factionId: _gp2,
-        ),
-        isFalse,
-      );
-    });
-
-    test('true when only a non-first invadable province is faction-owned', () {
-      // The .any short-circuit must still find the faction owner that is not
-      // the first scanned entry (other GP first, target second) -> true.
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      expect(
-        factionOwnsInvadableOldWorldProvince(
-          snapshot: snapshot,
-          provinceOwner: const {pA: _gp3, pB: _gp2},
-          factionId: _gp2,
-        ),
-        isTrue,
-      );
-    });
-
-    test('agrees with the inline scan it replaces (equivalence)', () {
-      final snapshot = _snapshotWithInvadable([pA, pB]);
-      for (final owner in <Map<String, String>>[
-        const {},
-        const {pA: _gp2},
-        const {pA: _gp3, pB: _gp2},
-        const {pA: _gp3, pB: _minor1},
-      ]) {
-        expect(
-          factionOwnsInvadableOldWorldProvince(
-            snapshot: snapshot,
-            provinceOwner: owner,
-            factionId: _gp2,
-          ),
-          snapshot.conquest.invadableProvinceIdsSorted.any(
-            (pid) => owner[pid] == _gp2,
-          ),
-          reason: 'mismatch for provinceOwner=$owner',
-        );
-      }
-    });
-  });
-
-  group('addInvadableProvinceMinorOwnersNotAtWar (Refs #3717)', () {
-    const String pA = 'provA';
-    const String pB = 'provB';
-    const String pC = 'provC';
-    const String _minor2 = 'minor2';
-
-    Game gameWithTwoMinors() => Game(
-      id: 'g-3717-minor-collector',
-      worldState: WorldState(
-        turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-        oldWorld: const RegionData(provinces: []),
-        newWorld: const RegionData(provinces: []),
-      ),
-      players: const [Player(id: _gp1, displayName: 'GP1', isHuman: false)],
-      minorNations: const [
-        MinorNation(id: _minor1, displayName: 'Minor1'),
-        MinorNation(id: _minor2, displayName: 'Minor2'),
-      ],
-      tribes: const [Tribe(id: _tribe1, displayName: 'Tribe1')],
-    );
-
-    AIWorldSnapshot snapshot(List<String> invadable, List<String> atWarWith) =>
-        AIWorldSnapshot(
-          playerId: _gp1,
-          threats: ThreatSummary(atWarWith: atWarWith),
-          opportunities: const OpportunitySummary(),
-          conquest: ConquestSummary(invadableProvinceIdsSorted: invadable),
-          economy: const EconomySummary(),
-          relations: const {},
-        );
-
-    test('collects minor owners of invadable provinces not already at war', () {
-      final into = <String>{};
-      addInvadableProvinceMinorOwnersNotAtWar(
-        game: gameWithTwoMinors(),
-        snapshot: snapshot([pA, pB], const []),
-        provinceOwner: const {pA: _minor1, pB: _minor2},
-        into: into,
-      );
-      expect(into, <String>{_minor1, _minor2});
-    });
-
-    test('skips minors already at war', () {
-      final into = <String>{};
-      addInvadableProvinceMinorOwnersNotAtWar(
-        game: gameWithTwoMinors(),
-        snapshot: snapshot([pA, pB], [_minor1]),
-        provinceOwner: const {pA: _minor1, pB: _minor2},
-        into: into,
-      );
-      expect(into, <String>{_minor2});
-    });
-
-    test('skips Great-Power, tribe, and unowned invadable provinces', () {
-      final into = <String>{};
-      addInvadableProvinceMinorOwnersNotAtWar(
-        game: gameWithTwoMinors(),
-        snapshot: snapshot([pA, pB, pC], const []),
-        // pA: GP, pB: tribe, pC: absent from the owner map (null lookup).
-        provinceOwner: const {pA: _gp1, pB: _tribe1},
-        into: into,
-      );
-      expect(into, isEmpty);
-    });
-
-    test('adds nothing when there are no invadable provinces', () {
-      final into = <String>{};
-      addInvadableProvinceMinorOwnersNotAtWar(
-        game: gameWithTwoMinors(),
-        snapshot: snapshot(const [], const []),
-        provinceOwner: const {pA: _minor1},
-        into: into,
-      );
-      expect(into, isEmpty);
-    });
-
-    test('preserves pre-seeded entries and de-duplicates via set', () {
-      // Mirrors the plateau decider seeding adjacent-owner candidates first.
-      final into = <String>{_minor1};
-      addInvadableProvinceMinorOwnersNotAtWar(
-        game: gameWithTwoMinors(),
-        snapshot: snapshot([pA, pB], const []),
-        provinceOwner: const {pA: _minor1, pB: _minor2},
-        into: into,
-      );
-      expect(into, <String>{_minor1, _minor2});
-    });
-
-    test('agrees with the inline collector loop it replaces (equivalence)', () {
-      final game = gameWithTwoMinors();
-      for (final atWar in <List<String>>[
-        const [],
-        [_minor1],
-        [_minor1, _minor2],
-      ]) {
-        for (final owner in <Map<String, String>>[
-          const {},
-          const {pA: _minor1, pB: _minor2},
-          const {pA: _gp1, pB: _minor2},
-          const {pA: _tribe1, pB: _minor1},
-        ]) {
-          final snap = snapshot([pA, pB], atWar);
-          final viaHelper = <String>{};
-          addInvadableProvinceMinorOwnersNotAtWar(
-            game: game,
-            snapshot: snap,
-            provinceOwner: owner,
-            into: viaHelper,
-          );
-          final viaInline = <String>{};
-          for (final pid in snap.conquest.invadableProvinceIdsSorted) {
-            final o = owner[pid];
-            if (o == null ||
-                !isMinorFaction(game, o) ||
-                snap.threats.atWarWith.contains(o)) {
-              continue;
-            }
-            viaInline.add(o);
-          }
-          expect(
-            viaHelper,
-            viaInline,
-            reason: 'mismatch for atWar=$atWar owner=$owner',
-          );
-        }
-      }
-    });
-  });
 }
